refactor(sql-playground): move service to feature folder
- Move sql-playground.service.ts to features/sql-playground/services/
- Update 5 import statements (components, effects, dialogs)
- Update service's model imports to use relative paths (../models/)
- Keep providedIn: 'root' for global DI availability
- Delete old service/sql-playground.service.ts
- No functional changes

Step 4 of 6
